removed DisposeModelAndConsumers will do it in another PR
